Move BseController from the Api::V1 namespace
